Http_mongodb/requirements.txt
fastapi==0.115.14
motor==3.7.1
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
python-dotenv==1.1.1
uvicorn==0.35.0